Lower V3AuthVotingInterval back to upstream defaults
Why? Because this makes the Tor network bootstrap way faster (3 times on my
machine).

Why might it OK? This reverts 38eacd13ae681cb03049c5f90db505680c9615d6, where we
bumped that value from 10 to 180. Around the same time we did that, Tor folks
bumped that value from 10 to 20, for similar reasons. They did not touch it
since, so it seems it works for them, and there's a chance it works for us too.
And who knows, perhaps the root cause of the problem was in the tor client,
and got fixed since.